reorganized to make sure .java files matched .h files
