+++ contrib/platform/src/com/sun/jna/platform/win32/WinUser.java
@@ -1559,4 +1559,56 @@ public interface WinUser extends StdCallLibrary, WinDef {
             return "hDevice=" + hDevice + ", dwType=" + dwType;
         }
     }
+
+	/**
+	 * An application-defined callback (or hook) function that the system calls
+	 * in response to events generated by an accessible object.<br>
+	 * The hook function processes the event notifications as required.<br>
+	 * Clients install the hook function and request specific types of event
+	 * notifications by calling SetWinEventHook.<br>
+	 * The WINEVENTPROC type defines a pointer to this callback function.
+	 * WinEventProc is a placeholder for the application-defined function name.
+	 */
+	interface WinEventProc extends Callback {
+		/**
+		 * @param hWinEventHook
+		 *            Type: HWINEVENTHOOK<br>
+		 *            Handle to an event hook function.<br>
+		 *            This value is returned by SetWinEventHook when the hook
+		 *            function is installed and is specific to each instance of
+		 *            the hook function.
+		 * @param event
+		 *            Type: DWORD<br>
+		 *            Specifies the event that occurred.<br>
+		 *            This value is one of the event constants.
+		 * @param hwnd
+		 *            Type: HWND<br>
+		 *            Handle to the window that generates the event, or NULL if
+		 *            no window is associated with the event.<br>
+		 *            For example, the mouse pointer is not associated with a
+		 *            window.
+		 * @param idObject
+		 *            Type: LONG<br>
+		 *            Identifies the object associated with the event.<br>
+		 *            This is one of the object identifiers or a custom object
+		 *            ID.
+		 * @param idChild
+		 *            Type: LONG<br>
+		 *            Identifies whether the event was triggered by an object or
+		 *            a child element of the object.<br>
+		 *            If this value is CHILDID_SELF, the event was triggered by
+		 *            the object; otherwise, this value is the child ID of the
+		 *            element that triggered the event.
+		 * @param dwEventThread
+		 *            Type: DWORD<br>
+		 *            Identifies the thread that generated the event, or the
+		 *            thread that owns the current window.
+		 * @param dwmsEventTime
+		 *            Type: DWORD<br>
+		 *            Specifies the time, in milliseconds, that the event was
+		 *            generated.
+		 */
+		void callback(HANDLE hWinEventHook, DWORD event, HWND hwnd, LONG idObject, LONG idChild, DWORD dwEventThread,
+				DWORD dwmsEventTime);
+	}
 }
